modify database to store geneated files; modify startJob to avoid reexecute workflow; version 1.2

+++ src/main/java/cz/zcu/kiv/server/scheduler/Manager.java
@@ -43,7 +43,7 @@ public class Manager {
     }
 
     //New Job arrives
-    public long addJob(Job job) throws SQLException {
+    public synchronized long addJob(Job job) throws SQLException {
         //Add job to database as a pending job
         job = Jobs.addJob(job);
 
@@ -58,6 +58,13 @@ public class Manager {
     //Thread is available, start execution
     public synchronized void startJob(Job job){
         CURRENT_THREAD_COUNT++;
+
+        job.setStatus(Status.RUNNING);
+        try {
+            Jobs.updateJob(job);
+        } catch (SQLException e) {
+            logger.error(e);
+        }
         JobThread jobThread=new JobThread(job){
             @Override
             public void onJobCompleted(){
@@ -72,7 +79,7 @@ public class Manager {
     }
 
     //Indicate thread is available and schedule pending job
-    public void endJob() {
+    public synchronized void endJob() {
 
         CURRENT_THREAD_COUNT--;
         try {
@@ -80,8 +87,11 @@ public class Manager {
             List<Job> pendingJobs = Jobs.getJobsByStatus(Status.WAITING.name());
             if(!pendingJobs.isEmpty()){
 
-                //check if thread is available and schedule job
-                startJob(pendingJobs.get(0));
+                // check if thread is available and schedule job
+                Job job = pendingJobs.get(0);
+                job.setStatus(Status.RUNNING);
+                Jobs.updateJob(job);
+                startJob(job);
             }
         } catch (SQLException e) {
             e.printStackTrace();
@@ -110,17 +120,22 @@ public class Manager {
     public void clearJobs(String email) throws SQLException {
         List<Job> jobs = Jobs.getNotRunningOrWaitingJobs(email);
         Jobs.clearJobs(email);
+
+        if (jobs == null || jobs.size() == 0) return;
+
+        // all the jobs belongs to the owner = this email
+        String outputFolder = GENERATED_FILES_FOLDER + email;
         for(Job item : jobs) {
+            // delete all the jsonFiles(in WORKING_DIRECTORY ) the owner has
             FileUtils.deleteQuietly(new java.io.File(item.getWorkflowOutputFile()));
-        }
 
-        File generatedFolder = new File(GENERATED_FILES_FOLDER);
-        String[] generatedFileNames= generatedFolder.list();
-
-        if(generatedFileNames == null) return;
-        for(String fileStr:  generatedFileNames){
-            FileUtils.deleteQuietly(new File(GENERATED_FILES_FOLDER+fileStr));
-            logger.info("delete generated file:"+fileStr);
+            // delete all the generatedFiles(in GENERATED_FILES_FOLDER ) of this job the owner has
+            List<String> files = item.getGeneratedFilesList();
+            if ( files == null || files.size() == 0) continue;
+            for(String generatedFile: files){
+                logger.info("delete " + email + "'s " + generatedFile);
+                FileUtils.deleteQuietly(new File(outputFolder + File.separator + generatedFile ));
+            }
         }
 
     }
